Plot confusion matrix as a heatmap Creates a heatmap visualization of a confusion matrix. Optionally saves the plot to a file.

Plot confusion matrix as a heatmap Creates a heatmap visualization of a confusion matrix. Optionally saves the plot to a file.

Usage

plot_confusion_matrix(
  cm,
  path_saved = NULL,
  color_low = "#f1f1f1",
  color_high = "#1f77b4",
  text_color = "white",
  text_size = 8,
  title = "Confusion Matrix Heatmap",
  width = NULL,
  height = NULL
)

Arguments

cm

A caret confusionMatrix object

path_saved

File path to save the plot. If NULL, plot is only returned but not saved.

color_low

Color for low values in gradient. Default: "#f1f1f1" (light gray)

color_high

Color for high values in gradient. Default: "#1f77b4" (blue)

text_color

Color for the text labels. Default: "white"

text_size

Size for the text labels. Default: 8

title

Plot title. Default: "Confusion Matrix Heatmap"

width

Plot width in inches when saving. Default: 8

height

Plot height in inches when saving. Default: 7

Value

Invisibly returns the ggplot object.

Examples

plot_confusion_matrix(cm)
#> Error in plot_confusion_matrix(cm): The 'cm' parameter must be a confusionMatrix object from the caret package
plot_confusion_matrix(cm, path_saved = "cm_plot.png")
#> Error in plot_confusion_matrix(cm, path_saved = "cm_plot.png"): The 'cm' parameter must be a confusionMatrix object from the caret package